fixed: razers2 VS7 error, some warnings removed
git-svn-id: http://svn.seqan.de/seqan/trunk@6212 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/razers3/paramChooser.h
@@ -64,7 +64,7 @@ int getDir(TPath path, TFilenameString &files)
 
 struct ParamChooserOptions
 {
-	typedef float TFloat;
+	typedef double TFloat;
 	unsigned minThreshold;					// minimum value for threshold parameter 
 	unsigned maxWeight;                                           // maximum value of q
 	bool chooseOneGappedOnly;      // choose onegapped (or ungapped) shape (discard all other gapped shapes)
